Move noise generation into math sub-package
