+++ gym_bhsl/envs/bhsl_env.py
@@ -8,7 +8,7 @@ import gym
 from gym import error, spaces
 from gym.utils import seeding
 
-from gym_bhsl.envs.noise import OUNoise
+from gym_bhsl.math.noise import OUNoise
 
 
 class Action(Enum):
